Modifying default cw lambda.
This lambda function is a modification of the default lambda that is selected when you choose to stream cloudwatch logs to elasticsearch.

This will allow you to stream multiple cloudwatch log groups to a single elasticsearch cluster. The default cloudwatch -> elasticsearch lambda function does not account for multiple indexes. Therefore if you do not modify the index= you will receive errors and be unable  to publish logs to elasticsearch.